requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
grpcio==1.74.0
grpcio-tools==1.74.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
